Update vue.md
Change where the machine definition is to better represent the decoupling between the machine and Vue (like in the React guide).
Add a context property to the machine definition.
Add a reference to the context to explain that it can be used as a data store for the app. 
Explain how to watch for context (data) changes.